Properly init commands. Move admin commands into their own subdir
