+++ commands/command.js
@@ -4,7 +4,10 @@ module.exports = class Command {
 
   constructor(name) {
     this.name = name
-    db.one('SELECT * FROM commands WHERE name = $1', [name], row => {
+  }
+
+  async init() {
+    await db.one('SELECT * FROM commands WHERE name = $1', [this.name], row => {
       this.mounted = row.mounted
       this.admin = row.admin
     })
